rename auth files to authentication
