+++ devU-api/src/authentication/login.developer/login.developer.controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response, NextFunction } from 'express'
 
 import UserService from '../../entities/user/user.service'
-import AuthService from '../auth.service'
+import AuthService from '../authentication.service'
 
 import { GenericResponse } from '../../utils/apiResponse.utils'
 import { refreshCookieOptions } from '../../utils/cookie.utils'
