+++ devU-api/src/authentication/login/login.controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response, NextFunction } from 'express'
 
 import UserService from '../../entities/user/user.service'
-import AuthService from '../auth.service'
+import AuthService from '../authentication.service'
 import ProviderService from '../../provider/provider.service'
 
 import { Unauthorized } from '../../utils/apiResponse.utils'
