+++ devU-api/src/authentication/login/tests/login.controller.test.ts
@@ -5,7 +5,7 @@ import controller from '../login.controller'
 import UserModel from '../../../entities/user/user.model'
 
 import UserService from '../../../entities/user/user.service'
-import AuthService from '../../auth.service'
+import AuthService from '../../authentication.service'
 import ProviderService from '../../../provider/provider.service'
 
 import Testing from '../../../utils/testing.utils'
